# improved intersection performance